TRIK feature: Add a metric system (#1959)
* Add a metric system
* Implicitly set default unit in twoDModelWidget
* Change precision for Pixels in Cm spinbox
* Fix clazy warnings
* Fix Vera issues
* Fix all clazy warnings
---------


---------
This `pull request` is one of many proposed ones and is the first step towards updating `twoDModel`. This `pull request` does not address issues related to the `physical parameters` of `world items` or (for example) the ability to resize objects (unless this is already implemented in the previous version).

1. Add the `sizeUnit`(`mm`, `m`, `cm`) field to the <settings> tag in the `WorldModel xml` to specify units of measurement when working with items.
2. Thus, the `coordinates`, `starting position`, and `sizes` for all objects specified in `WorldModel xml` (`robot`, `sensor`, `ball`, `region`, ...) will be interpreted based on the value of the 'sizeUnit` field (pixels by default).
3. The user interface for selecting units of measurement is located in the panel on the bottom right. Changing the unit of measurement implies changing the `Grid` (as well as loading the `WorldModel`) and also affects the further serialization of objects.

<img width="1481" height="913" alt="image" src="https://github.com/user-attachments/assets/f9f6286c-7e0b-46e5-a402-8ea3355f874d" />

4. This functionality is only used for `TRIK twoDModel`. When using the `sizeUnit` field with older `TRIK Studio` versions, all values will be interpreted by default (`pixels`).
5. Add unit tests for `WorldModel` deserialization.
6. Also, in the future, it is necessary to consider the possibility of changing parameters in the `UI` with `PopupItem` in the current units of measurement, as well as making the `Pixels in cm` field editable (part of the next pull request).

+++ plugins/robots/common/kitBase/include/kitBase/robotModel/portInfo.h
@@ -97,7 +97,7 @@ inline bool operator !=(const PortInfo &left, const PortInfo &right)
 	return !(left == right);
 }
 
-inline uint qHash(const PortInfo &key)
+inline size_t qHash(const PortInfo &key)
 {
 	return qHash(key.name()) ^ qHash(QString(key.direction() == input ? "input" : "output"));
 }
